refactor(Event): 클래스 구조 조정 및 디렉토리 구조 조정
기존에 Event를 상속받는 Discount 인터페이스가 존재했던 반면에, Discount는 그 자체로 할인이라는 인터페이스로 혼자 남도록 한다. Event는 이벤트를 통해 활성화되는 할인 등에 사용되므로, 이에 BenefitDetail 앞에 Event를 붙인다.

Event 관련 클래스들은 Event 하위로 이동하고, 지금까지의 할인, 앞으로의 증정 이벤트는 하위 패키지에 생성한다.
