+++ src/main/java/christmas/domain/event/Event.java
@@ -1,4 +1,7 @@
 package christmas.domain.event;
 
+import christmas.domain.menu.OrderMenu;
+
 public interface Event {
+    EventBenefitDetail getBenefitDetail(OrderMenu orderMenu);
 }
