+++ src/main/java/christmas/domain/event/discount/Discount.java
@@ -1,9 +1,8 @@
 package christmas.domain.event.discount;
 
 import christmas.domain.Money;
-import christmas.domain.event.Event;
 import christmas.domain.menu.OrderMenu;
 
-public interface Discount extends Event {
+public interface Discount {
     Money apply(OrderMenu orderMenu);
 }
