+++ src/test/java/christmas/domain/event/discount/EventDiscountTest.java
@@ -1,6 +1,8 @@
 package christmas.domain.event.discount;
 
 import christmas.domain.Money;
+import christmas.domain.event.EventBenefitDetail;
+import christmas.domain.event.EventDiscount;
 import christmas.domain.menu.Menu;
 import christmas.domain.menu.MenuAmount;
 import christmas.domain.menu.OrderMenu;
@@ -23,8 +25,8 @@ class EventDiscountTest {
         EventDiscount alwaysDiscount = new EventDiscount("무조건 천원 할인", (orders -> Money.of(1_000L)), (date) -> true);
 
         // when
-        DiscountDetail discountDetail = alwaysDiscount.apply(orderMenu);
+        EventBenefitDetail eventBenefitDetail = alwaysDiscount.getBenefitDetail(orderMenu);
         // then
-        Assertions.assertThat(discountDetail.toString()).isEqualTo("무조건 천원 할인: 1,000원");
+        Assertions.assertThat(eventBenefitDetail.toString()).isEqualTo("무조건 천원 할인: 1,000원");
     }
 }
